refactor cpu preamble
